+++ .github/scripts/pr_triage.py
@@ -1,311 +1,35 @@
-"""Decides what to do with a pull request on open/reopen, and does the writes.
+"""Decides what to do with a pull request on open/reopen, and does the
+writes: main(), the event handlers, the apply/orchestration functions that
+mix a decision with I/O, and the step summary.
 
-The decision (decide) is pure: author login + current draft state + the loaded
-teams mapping in, an assign/draft/board verdict out. No network, no GitHub
-API, no environment reads. Everything below main() is I/O built on top of it.
+The pure verdicts live in pr_triage_decide (no network, no GitHub API, no
+environment reads, no clock); every HTTP call lives in pr_triage_github. This
+module wires the two together and is the one the workflow runs directly, so
+the script directory is put on sys.path before either sibling import.
 """
 
 import datetime
 import json
 import os
-import re
 import sys
-import urllib.error
-import urllib.request
 
-BOARD_PROJECT_ID = "PVT_kwDOABnCXs4AgIEx"
-STATUS_FIELD_ID = "PVTSSF_lADOABnCXs4AgIExzgVUF6A"
-TEAM_FIELD_ID = "PVTSSF_lADOABnCXs4AgIExzgbkzoM"
-SPRINT_FIELD_ID = "PVTIF_lADOABnCXs4AgIExzgavkek"
-STATUS_WORKING = "Working"
-STATUS_REVIEW = "Review"
-UNKNOWN_AUTHOR_TEAM = "Dev"
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 
-GITHUB_API_URL = "https://api.github.com"
-GITHUB_GRAPHQL_URL = "https://api.github.com/graphql"
-HTTP_TIMEOUT_SECONDS = 30
-
-
-class GraphqlError(Exception):
-    pass
-
-
-class AssignmentDroppedError(Exception):
-    pass
-
-
-def check_assignment_succeeded(author_login, assign_result):
-    assignee_logins = [assignee["login"] for assignee in assign_result.get("assignees", [])]
-    if author_login not in assignee_logins:
-        raise AssignmentDroppedError(
-            f"GitHub did not assign '{author_login}' (assignees after the call: {assignee_logins}). "
-            f"The login is probably no longer valid for assignment; .github/pr-triage-teams.yml is probably stale."
-        )
-
-
-def is_blank(value):
-    return not value or not value.strip()
-
-
-def team_for(author_login, mapping):
-    team = mapping.get(author_login)
-    return None if is_blank(team) else team
-
-
-def decide(author_login, is_draft, mapping):
-    team = team_for(author_login, mapping)
-    if team is None:
-        return {"known": False, "force_draft": False, "team": UNKNOWN_AUTHOR_TEAM}
-    return {"known": True, "force_draft": not is_draft, "team": team}
-
-
-def target_status_from_draft_state(is_draft):
-    return STATUS_WORKING if is_draft else STATUS_REVIEW
+import pr_triage_decide
+import pr_triage_github
 
 
 def current_date():
     return datetime.date.today()
 
 
-def find_current_iteration(iterations, today):
-    for iteration in iterations:
-        start = datetime.date.fromisoformat(iteration["startDate"])
-        end = start + datetime.timedelta(days=iteration["duration"])
-        if start <= today < end:
-            return iteration
-    return None
-
-
-STATUS_MANAGED_VALUES = (STATUS_WORKING, STATUS_REVIEW)
-
-
-def decide_status(is_draft, current_status):
-    target = target_status_from_draft_state(is_draft)
-    if is_blank(current_status):
-        return target
-    current = strip_emoji_prefix(current_status)
-    if current == target:
-        return None
-    if current in STATUS_MANAGED_VALUES:
-        return target
-    return None
-
-
-def decide_board_update(is_draft, current_status, current_team, mapped_team, current_sprint=None, current_sprint_title=None):
-    return {
-        "status": decide_status(is_draft, current_status),
-        "team": mapped_team if is_blank(current_team) else None,
-        "sprint": current_sprint_title if (current_sprint_title and is_blank(current_sprint)) else None,
-    }
-
-
-def keyword_closing_issues(all_closing_issues, user_linked_issues):
-    """Aim 6 redirects only on a KEYWORD-derived closing reference
-    (close/closes/closed/fix/fixes/fixed/resolve/resolves/resolved in the PR
-    body) [owner ruling, narrowing the original "closingIssuesReferences
-    reports it, so it counts" reading]. An issue linked only by hand through
-    GitHub's Development sidebar, with no keyword in the body, does NOT
-    redirect -- `Addresses: <url>` is a real habit in this repo, used
-    deliberately to link without auto-closing.
-
-    GitHub's closingIssuesReferences field reports both kinds together;
-    userLinkedOnly=true reports only the hand-linked ones. Keyword-derived is
-    therefore `all_closing_issues` minus `user_linked_issues`, matched by
-    node id -- verified against two real PRs of each shape, see
-    notes/github-facts.md §8."""
-    user_linked_ids = {issue["id"] for issue in user_linked_issues}
-    return [issue for issue in all_closing_issues if issue["id"] not in user_linked_ids]
-
-
-def decide_redirect_to_issues(closing_issues):
-    """Aim 6: a PR that closes at least one issue gets no card of its own --
-    every linked issue's card is written instead. An empty list changes
-    nothing; the PR is boarded exactly as it always was."""
-    return bool(closing_issues)
-
-
-def decide_remove_pr_card(closing_issues, existing_pr_item):
-    """The story's one removal: unconditional, and narrow. True exactly when
-    the PR closes at least one issue AND the PR already has an item on board
-    15 -- never guarded by that item's Status or fields, and never true for
-    an issue's item, which this is never called with. "Removal" is
-    board-only: the card comes off board 15, nothing about the PR itself
-    changes.
-
-    Deliberately action-blind: this function does not know synchronize is
-    excluded from removal. That exclusion lives in the one call site,
-    apply_board_write's `action not in PR_CARD_REMOVAL_ACTIONS` guard --
-    apply_redirected_board_write (and this function) are never reached for
-    synchronize at all. If a second caller is ever added, that guard has to
-    move or be duplicated; it does not enforce itself here."""
-    return decide_redirect_to_issues(closing_issues) and existing_pr_item is not None
-
-
-def decide_issue_card_update(action, is_draft, existing_issue_item, mapped_team, current_sprint_title):
-    """Pure decision for ONE linked issue's card. existing_issue_item is None
-    when the issue has no item yet on board 15.
-
-    - `synchronize` never touches a linked issue's card at all -- existing or
-      missing, this returns None.
-    - `edited` never touches an EXISTING issue card (Status is not moved by a
-      body edit), but still adds and fills a missing one, because a PR with
-      no card and an issue with no card is the one outcome no aim tolerates.
-      Status is left unwritten on the card it adds -- only opened and the
-      three transitions ever move a linked issue's Status.
-    - `opened` and the three transitions set Status only on an existing card
-      (Team and Sprint stay exactly as a person left them), and fill all
-      three -- Status, Team, Sprint -- on a card they add.
-
-    Returns None for "write nothing", otherwise a fields dict shaped like
-    resolve_and_apply_board_fields expects: {"status", "team", "sprint"},
-    each either a target value or None meaning "leave as-is"."""
-    if existing_issue_item is not None:
-        if action not in ISSUE_STATUS_ACTIONS:
-            return None
-        return {"status": decide_status(is_draft, existing_issue_item["status"]), "team": None, "sprint": None}
-
-    if action not in ISSUE_STATUS_ACTIONS and action not in ISSUE_ADD_ONLY_ACTIONS:
-        return None
-    status = target_status_from_draft_state(is_draft) if action in ISSUE_STATUS_ACTIONS else None
-    return {"status": status, "team": mapped_team, "sprint": current_sprint_title}
-
-
-def strip_emoji_prefix(name):
-    return re.sub(r"^[^\w]+", "", name).strip()
-
-
-def find_option_id_by_name(options, target_name, strip_emoji):
-    for option in options:
-        candidate = strip_emoji_prefix(option["name"]) if strip_emoji else option["name"]
-        if candidate == target_name:
-            return option["id"]
-    available = ", ".join(f'{o["name"]!r} ({o["id"]})' for o in options)
-    raise ValueError(
-        f"'{target_name}' does not match any live board option. Available options: {available}"
-    )
-
-
-def parse_teams_entries(text):
-    entries = []
-    in_teams_block = False
-    for raw_line in text.splitlines():
-        line = raw_line.split("#", 1)[0].rstrip()
-        if not line.strip():
-            continue
-        if line.strip() == "teams:":
-            in_teams_block = True
-            continue
-        if not in_teams_block:
-            continue
-        if not line.startswith((" ", "\t")):
-            in_teams_block = False
-            continue
-        key, _, value = line.strip().partition(":")
-        entries.append((key.strip(), value.strip()))
-    return entries
-
-
-def parse_teams_mapping(text):
-    return dict(parse_teams_entries(text))
-
-
 def load_teams_mapping(path):
     with open(path, encoding="utf-8") as handle:
-        return parse_teams_mapping(handle.read())
+        return pr_triage_decide.parse_teams_mapping(handle.read())
 
 
 def mapping_file_path(pr_triage_file):
     return os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(pr_triage_file))), "pr-triage-teams.yml")
-
-
-def http_request(url, token, method="GET", body=None):
-    data = json.dumps(body).encode("utf-8") if body is not None else None
-    request = urllib.request.Request(url, data=data, method=method)
-    request.add_header("Authorization", f"Bearer {token}")
-    request.add_header("Accept", "application/vnd.github+json")
-    if data is not None:
-        request.add_header("Content-Type", "application/json")
-    try:
-        with urllib.request.urlopen(request, timeout=HTTP_TIMEOUT_SECONDS) as response:
-            return json.loads(response.read().decode("utf-8"))
-    except urllib.error.HTTPError as error:
-        print(f"GitHub API error {error.code} for {method} {url}: {error.read().decode('utf-8')}")
-        raise
-
-
-def graphql_request(query, variables, token):
-    result = http_request(
-        GITHUB_GRAPHQL_URL,
-        token,
-        method="POST",
-        body={"query": query, "variables": variables},
-    )
-    errors = result.get("errors")
-    if errors:
-        raise GraphqlError(f"GraphQL request returned errors: {errors}")
-    return result
-
-
-def require_node(result, resolved_id):
-    node = result["data"]["node"]
-    if node is None:
-        raise GraphqlError(
-            f"GraphQL id '{resolved_id}' resolved to no node (data.node is null). "
-            f"The id is probably stale, e.g. a board field that was deleted and recreated."
-        )
-    return node
-
-
-def assign_author(repo, pr_number, author_login, token):
-    return http_request(
-        f"{GITHUB_API_URL}/repos/{repo}/issues/{pr_number}/assignees",
-        token,
-        method="POST",
-        body={"assignees": [author_login]},
-    )
-
-
-def convert_pr_to_draft(pr_node_id, token):
-    query = """
-    mutation($pullRequestId: ID!) {
-      convertPullRequestToDraft(input: { pullRequestId: $pullRequestId }) {
-        pullRequest { id isDraft }
-      }
-    }
-    """
-    return graphql_request(query, {"pullRequestId": pr_node_id}, token)
-
-
-def add_item_to_project(pr_node_id, token):
-    query = """
-    mutation($projectId: ID!, $contentId: ID!) {
-      addProjectV2ItemById(input: { projectId: $projectId, contentId: $contentId }) {
-        item { id }
-      }
-    }
-    """
-    result = graphql_request(
-        query, {"projectId": BOARD_PROJECT_ID, "contentId": pr_node_id}, token
-    )
-    return result["data"]["addProjectV2ItemById"]["item"]["id"]
-
-
-def remove_item_from_board(item_id, token):
-    """The story's one removal (aim 6): a PR's own CARD on board 15, removed
-    unconditionally the moment the PR closes an issue and already has one --
-    no guard on its Status or fields. This removes the board item only,
-    nothing about the pull request itself; the GraphQL mutation is still
-    named deleteProjectV2Item (GitHub's name, not ours). Never call this with
-    an issue's item id; an issue's card is never removed by anything in this
-    workflow."""
-    query = """
-    mutation($projectId: ID!, $itemId: ID!) {
-      deleteProjectV2Item(input: { projectId: $projectId, itemId: $itemId }) {
-        deletedItemId
-      }
-    }
-    """
-    return graphql_request(query, {"projectId": BOARD_PROJECT_ID, "itemId": item_id}, token)
 
 
 def find_board_item_for_pr(content_id, token):
@@ -317,7 +41,10 @@ def find_board_item_for_pr(content_id, token):
     riding both the unfiltered and userLinkedOnly variants on this same query
     costs no extra round trip. keyword_closing_issues does the split -- see
     notes/github-facts.md §8. Returns {"item": item-or-None, "closing_issues":
-    [{"id", "number"}, ...]} -- closing_issues is always [] for an Issue node."""
+    [{"id", "number"}, ...]} -- closing_issues is always [] for an Issue node.
+
+    Lives here rather than in pr_triage_github because it needs
+    keyword_closing_issues, and pr_triage_github imports nothing of ours."""
     query = """
     query($contentId: ID!) {
       node(id: $contentId) {
@@ -351,11 +78,11 @@ def find_board_item_for_pr(content_id, token):
       }
     }
     """
-    result = graphql_request(query, {"contentId": content_id}, token)
-    node = require_node(result, content_id)
+    result = pr_triage_github.graphql_request(query, {"contentId": content_id}, token)
+    node = pr_triage_github.require_node(result, content_id)
     item = None
     for candidate in node["projectItems"]["nodes"]:
-        if candidate["project"]["id"] == BOARD_PROJECT_ID:
+        if candidate["project"]["id"] == pr_triage_github.BOARD_PROJECT_ID:
             status_value = candidate["status"]
             team_value = candidate["team"]
             sprint_value = candidate["sprint"]
@@ -370,92 +97,8 @@ def find_board_item_for_pr(content_id, token):
     all_closing_issues = closing_issues_field["nodes"] if closing_issues_field else []
     user_linked_field = node.get("userLinkedClosingIssues")
     user_linked_issues = user_linked_field["nodes"] if user_linked_field else []
-    closing_issues = keyword_closing_issues(all_closing_issues, user_linked_issues)
+    closing_issues = pr_triage_decide.keyword_closing_issues(all_closing_issues, user_linked_issues)
     return {"item": item, "closing_issues": closing_issues}
-
-
-def set_project_field_option(item_id, field_id, option_id, token):
-    query = """
-    mutation($projectId: ID!, $itemId: ID!, $fieldId: ID!, $optionId: String!) {
-      updateProjectV2ItemFieldValue(
-        input: {
-          projectId: $projectId
-          itemId: $itemId
-          fieldId: $fieldId
-          value: { singleSelectOptionId: $optionId }
-        }
-      ) {
-        projectV2Item { id }
-      }
-    }
-    """
-    return graphql_request(
-        query,
-        {
-            "projectId": BOARD_PROJECT_ID,
-            "itemId": item_id,
-            "fieldId": field_id,
-            "optionId": option_id,
-        },
-        token,
-    )
-
-
-def fetch_project_field_options(field_id, token):
-    query = """
-    query($fieldId: ID!) {
-      node(id: $fieldId) {
-        ... on ProjectV2SingleSelectField {
-          options { id name }
-        }
-      }
-    }
-    """
-    result = graphql_request(query, {"fieldId": field_id}, token)
-    return require_node(result, field_id)["options"]
-
-
-def set_project_field_iteration(item_id, field_id, iteration_id, token):
-    query = """
-    mutation($projectId: ID!, $itemId: ID!, $fieldId: ID!, $iterationId: String!) {
-      updateProjectV2ItemFieldValue(
-        input: {
-          projectId: $projectId
-          itemId: $itemId
-          fieldId: $fieldId
-          value: { iterationId: $iterationId }
-        }
-      ) {
-        projectV2Item { id }
-      }
-    }
-    """
-    return graphql_request(
-        query,
-        {
-            "projectId": BOARD_PROJECT_ID,
-            "itemId": item_id,
-            "fieldId": field_id,
-            "iterationId": iteration_id,
-        },
-        token,
-    )
-
-
-def fetch_project_iterations(field_id, token):
-    query = """
-    query($fieldId: ID!) {
-      node(id: $fieldId) {
-        ... on ProjectV2IterationField {
-          configuration {
-            iterations { id title startDate duration }
-          }
-        }
-      }
-    }
-    """
-    result = graphql_request(query, {"fieldId": field_id}, token)
-    return require_node(result, field_id)["configuration"]["iterations"]
 
 
 def write_step_summary(rows):
@@ -475,26 +118,9 @@ def write_step_summary(rows):
 # (iterationId, not singleSelectOptionId), which is why it is written separately
 # below rather than through SINGLE_SELECT_FIELD_SPECS; a real fourth field would
 # still touch this table, current_values, FIELD_LABELS, and the write loop below.
-TRANSITION_ACTIONS = ("ready_for_review", "converted_to_draft", "reopened")
-BOARD_UPDATE_ACTIONS = TRANSITION_ACTIONS + ("synchronize", "edited")
-
-# Aim 6, redirected to a linked issue's card instead of the PR's own: which
-# actions may set that card's Status, and which may only add+fill a missing
-# one without ever moving Status. See decide_issue_card_update.
-ISSUE_STATUS_ACTIONS = TRANSITION_ACTIONS + ("opened",)
-ISSUE_ADD_ONLY_ACTIONS = ("edited",)
-
-# The actions that remove the PR's own card when it closes an issue: opened,
-# edited, and the three transitions. NOT synchronize -- a push that removed
-# the PR's card while never adding the issue's would leave zero cards for
-# the same piece of work (owner ruling, review finding F1). For a closing
-# PR, synchronize therefore writes nothing anywhere: not the PR's card
-# (removal or otherwise), not any issue's.
-PR_CARD_REMOVAL_ACTIONS = ISSUE_STATUS_ACTIONS + ISSUE_ADD_ONLY_ACTIONS
-
 SINGLE_SELECT_FIELD_SPECS = (
-    ("status", STATUS_FIELD_ID, True, "Status"),
-    ("team", TEAM_FIELD_ID, False, "Team"),
+    ("status", pr_triage_github.STATUS_FIELD_ID, True, "Status"),
+    ("team", pr_triage_github.TEAM_FIELD_ID, False, "Team"),
 )
 FIELD_LABELS = {"status": "Status", "team": "Team", "sprint": "Sprint"}
 
@@ -504,8 +130,8 @@ def resolve_and_apply_board_fields(pr_node_id, existing_item, fields, current_va
     for key, field_id, strip_emoji, _label in SINGLE_SELECT_FIELD_SPECS:
         target_value = fields[key]
         if target_value is not None:
-            options = fetch_project_field_options(field_id, board_token)
-            resolved_single_select[key] = (field_id, find_option_id_by_name(options, target_value, strip_emoji=strip_emoji))
+            options = pr_triage_github.fetch_project_field_options(field_id, board_token)
+            resolved_single_select[key] = (field_id, pr_triage_decide.find_option_id_by_name(options, target_value, strip_emoji=strip_emoji))
 
     sprint_iteration_id = current_iteration["id"] if (fields["sprint"] is not None and current_iteration) else None
 
@@ -513,7 +139,7 @@ def resolve_and_apply_board_fields(pr_node_id, existing_item, fields, current_va
         item_id = existing_item["id"]
         summary_rows.append(("Board item", f"found existing item {item_id}"))
     else:
-        item_id = add_item_to_project(pr_node_id, board_token)
+        item_id = pr_triage_github.add_item_to_project(pr_node_id, board_token)
         summary_rows.append(("Board item", f"no existing item, added {item_id}"))
 
     for key, label in FIELD_LABELS.items():
@@ -521,19 +147,19 @@ def resolve_and_apply_board_fields(pr_node_id, existing_item, fields, current_va
         if target_value is None:
             summary_rows.append((label, f"left as-is ({current_values[key]!r})"))
         elif key == "sprint":
-            set_project_field_iteration(item_id, SPRINT_FIELD_ID, sprint_iteration_id, board_token)
+            pr_triage_github.set_project_field_iteration(item_id, pr_triage_github.SPRINT_FIELD_ID, sprint_iteration_id, board_token)
             summary_rows.append((f"{label} set", target_value))
         else:
             field_id, option_id = resolved_single_select[key]
-            set_project_field_option(item_id, field_id, option_id, board_token)
+            pr_triage_github.set_project_field_option(item_id, field_id, option_id, board_token)
             summary_rows.append((f"{label} set", target_value))
 
     return item_id
 
 
 def resolve_current_iteration(board_token):
-    iterations = fetch_project_iterations(SPRINT_FIELD_ID, board_token)
-    return find_current_iteration(iterations, current_date())
+    iterations = pr_triage_github.fetch_project_iterations(pr_triage_github.SPRINT_FIELD_ID, board_token)
+    return pr_triage_decide.find_current_iteration(iterations, current_date())
 
 
 def board_verdict(prefix, fields, current_values):
@@ -572,7 +198,7 @@ def apply_redirected_board_write(action, is_draft, mapped_team, pr_item, closing
     for issue_ref in closing_issues:
         issue_lookup = find_board_item_for_pr(issue_ref["id"], board_token)
         existing_issue_item = issue_lookup["item"]
-        fields = decide_issue_card_update(action, is_draft, existing_issue_item, mapped_team, current_sprint_title)
+        fields = pr_triage_decide.decide_issue_card_update(action, is_draft, existing_issue_item, mapped_team, current_sprint_title)
         if fields is None:
             summary_rows.append((f"Issue #{issue_ref['number']}", "left untouched"))
             continue
@@ -586,13 +212,13 @@ def apply_redirected_board_write(action, is_draft, mapped_team, pr_item, closing
         )
         summary_rows.append((f"Issue #{issue_ref['number']}", f"item {item_id}"))
 
-    if decide_remove_pr_card(closing_issues, pr_item):
+    if pr_triage_decide.decide_remove_pr_card(closing_issues, pr_item):
         was = f"Status {pr_item['status']!r}, Team {pr_item['team']!r}, Sprint {pr_item['sprint']!r}"
         # Recorded BEFORE the mutation runs -- this row is the one place a
         # person can recover the values from, so a timeout after GitHub has
         # already committed the removal must not lose it.
         summary_rows.append(("PR card removed", f"{pr_item['id']} (was: {was})"))
-        remove_item_from_board(pr_item["id"], board_token)
+        pr_triage_github.remove_item_from_board(pr_item["id"], board_token)
     else:
         summary_rows.append(("PR card", f"none to add or remove (closes {len(closing_issues)} linked issue(s))"))
 
@@ -609,8 +235,8 @@ def apply_board_write(action, pr_node_id, is_draft, mapped_team, board_token, su
     pr_item = lookup["item"]
     closing_issues = lookup["closing_issues"]
 
-    if decide_redirect_to_issues(closing_issues):
-        if action not in PR_CARD_REMOVAL_ACTIONS:
+    if pr_triage_decide.decide_redirect_to_issues(closing_issues):
+        if action not in pr_triage_decide.PR_CARD_REMOVAL_ACTIONS:
             _record_verdict(
                 summary_rows, insert_verdict_at, f"{verdict_prefix} -> closes {len(closing_issues)} issue(s), {action} writes nothing"
             )
@@ -627,11 +253,11 @@ def apply_board_write(action, pr_node_id, is_draft, mapped_team, board_token, su
         "sprint": pr_item["sprint"] if pr_item else None,
     }
     current_iteration = None
-    if is_blank(current_values["sprint"]):
+    if pr_triage_decide.is_blank(current_values["sprint"]):
         current_iteration = resolve_current_iteration(board_token)
     current_sprint_title = current_iteration["title"] if current_iteration else None
 
-    fields = decide_board_update(
+    fields = pr_triage_decide.decide_board_update(
         is_draft,
         current_values["status"],
         current_values["team"],
@@ -662,7 +288,7 @@ def handle_board_update(action, pull_request):
     try:
         mapping_path = mapping_file_path(__file__)
         mapping = load_teams_mapping(mapping_path)
-        mapped_team = team_for(author_login, mapping) or UNKNOWN_AUTHOR_TEAM
+        mapped_team = pr_triage_decide.team_for(author_login, mapping) or pr_triage_decide.UNKNOWN_AUTHOR_TEAM
 
         board_token = os.environ["PROJECT_BOARD_TOKEN"]
 
@@ -716,7 +342,7 @@ def handle_open_triage(action, pull_request, repo):
         mapping_path = mapping_file_path(__file__)
         mapping = load_teams_mapping(mapping_path)
 
-        decision = decide(author_login=author_login, is_draft=is_draft, mapping=mapping)
+        decision = pr_triage_decide.decide(author_login=author_login, is_draft=is_draft, mapping=mapping)
 
         board_token = os.environ["PROJECT_BOARD_TOKEN"]
 
@@ -724,8 +350,8 @@ def handle_open_triage(action, pull_request, repo):
         if decision["known"]:
             try:
                 github_token = os.environ["GITHUB_TOKEN"]
-                assign_result = assign_author(repo, pr_number, author_login, github_token)
-                check_assignment_succeeded(author_login, assign_result)
+                assign_result = pr_triage_github.assign_author(repo, pr_number, author_login, github_token)
+                pr_triage_decide.check_assignment_succeeded(author_login, assign_result)
                 summary_rows.append(("Assignee set", True))
             except Exception as error:
                 print(f"ERROR: assignment failed: {error}")
@@ -742,7 +368,7 @@ def handle_open_triage(action, pull_request, repo):
         actual_is_draft = is_draft
         if decision["force_draft"] and assignment_succeeded:
             try:
-                draft_result = convert_pr_to_draft(pr_node_id, board_token)
+                draft_result = pr_triage_github.convert_pr_to_draft(pr_node_id, board_token)
                 actual_is_draft = draft_result["data"]["convertPullRequestToDraft"]["pullRequest"]["isDraft"]
                 summary_rows.append(("Converted to draft", actual_is_draft))
             except Exception as error:
@@ -792,7 +418,7 @@ def main():
     action = event["action"]
     pull_request = event["pull_request"]
 
-    if action in BOARD_UPDATE_ACTIONS:
+    if action in pr_triage_decide.BOARD_UPDATE_ACTIONS:
         handle_board_update(action, pull_request)
         return
