+++ .github/workflows/pr-triage.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Validate pr-triage-teams.yml
         run: python3 .github/scripts/validate_pr_triage_teams.py
       - name: Run unit tests
-        run: python3 .github/scripts/test_pr_triage.py
+        run: python3 -m unittest discover -s .github/scripts
 
   triage:
     needs: validate
